refactor: 모듈 및 상수 분리
