+++ src/js/srcs/clearDisplay.js
@@ -1,0 +1,5 @@
+import { DISPLAY } from '../utils/DOM.js';
+
+export const clearDisplay = () => {
+  DISPLAY.innerText = '0';
+};
