+++ src/js/utils/DOM.js
@@ -1,0 +1,1 @@
+export const DISPLAY = document.querySelector('#total');
